chore: add first implementation
This commit adds basic functionality. The public API of the library
won't change much (if at all).

The binary has been implemented and can be used to generate a changelog
inside the current directory.

BREAKING CHANGE: The MSRV had to be bumped to allow for the `indoc`
dependency to compile.