refactor!: `generate_changelog` takes path instead of repository
This allows for easier call logic of extractors as operations can now be
performed on the directory itself instead of the repository instance.
The repository is opened internally instead.

The API change resulted in a new error variant `NoGitRepository` that
gets returned if `git2` wasn't able to open the repository in the passed
directory.

BREAKING CHANGE: The public method `generate_changelog` now takes `impl
AsRef<Path>` as an argument instead of a `git2::Repository` reference.

Closes #3